Conflicts:
	src/jquery.tokeninput.js

+++ src/jquery.tokeninput.js
@@ -1,6 +1,6 @@
 /*
  * jQuery Plugin: Tokenizing Autocomplete Text Entry
- * Version 1.4.2
+ * Version 1.5.0
  * Requires jQuery 1.6+
  *
  * Copyright (c) 2009 James Smith (http://loopj.com)
@@ -28,11 +28,14 @@ var DEFAULT_SETTINGS = {
     tokenDelimiter: ",",
     preventDuplicates: false,
     prePopulate: null,
+    processPrePopulate: false,
     animateDropdown: true,
     dontAdd: false,
     onResult: null,
     onAdd: null,
-    onDelete: null
+    onDelete: null,
+    idPrefix: "token-input-",
+    backspaceDeleteItem: true
 };
 
 // Default classes to use when theming
@@ -77,14 +80,36 @@ var KEY = {
     COMMA: 188
 };
 
+// Additional public (exposed) methods
+var methods = {
+    init: function(url_or_data_or_function, options) {
+        return this.each(function () {
+            $(this).data("tokenInputObject", new $.TokenList(this, url_or_data_or_function, options));
+        });
+    },
+    clear: function() {
+        this.data("tokenInputObject").clear();
+        return this;
+    },
+    add: function(item) {
+        this.data("tokenInputObject").add(item);
+        return this;
+    },
+    remove: function(item) {
+        this.data("tokenInputObject").remove(item);
+        return this;
+    }
+}
 
 // Expose the .tokenInput function to jQuery as a plugin
-$.fn.tokenInput = function (url_or_data, options) {
-    return this.each(function () {
-        new $.TokenList(this, url_or_data, options);
-    });
+$.fn.tokenInput = function (method) {
+    // Method calling and initialization logic
+    if(methods[method]) {
+        return methods[method].apply(this, Array.prototype.slice.call(arguments, 1));
+    } else {
+        return methods.init.apply(this, arguments);
+    }
 };
-
 
 // TokenList class for each input
 $.TokenList = function (input, url_or_data, options) {
@@ -144,6 +169,7 @@ $.TokenList = function (input, url_or_data, options) {
         .css({
             outline: "none"
         })
+        .attr("id", settings.idPrefix + input.id)
         .focus(function () {
             if (settings.tokenLimit === null || settings.tokenLimit !== token_count) {
               if(settings.permanentDropdown || settings.showAllResults) {
@@ -214,7 +240,7 @@ $.TokenList = function (input, url_or_data, options) {
                     previous_token = input_token.prev();
                     next_token = input_token.next();
 
-                    if(!$(this).val().length) {
+                    if(!$(this).val().length && settings.backspaceDeleteItem) {
                         if(selected_token) {
                             delete_token($(selected_token));
                             input_box.focus();
@@ -238,7 +264,7 @@ $.TokenList = function (input, url_or_data, options) {
                 case KEY.NUMPAD_ENTER:
                 case KEY.COMMA:
                   if(selected_dropdown_item) {
-                    add_token($(selected_dropdown_item));
+                    add_token($(selected_dropdown_item).data("tokeninput"));
                     input_box.focus();
                     return false;
                   }
@@ -348,18 +374,65 @@ $.TokenList = function (input, url_or_data, options) {
 
     // Pre-populate list if items exist
     hidden_input.val("");
-    li_data = settings.prePopulate || hidden_input.data("pre");
+    var li_data = settings.prePopulate || hidden_input.data("pre");
+    if(settings.processPrePopulate && $.isFunction(settings.onResult)) {
+        li_data = settings.onResult.call(hidden_input, li_data);
+    }
     if(li_data && li_data.length) {
         $.each(li_data, function (index, value) {
-            insert_token(value.id, value.name);
+            insert_token(value);
+            checkTokenLimit();
         });
     }
 
 
+    //
+    // Public functions
+    //
+
+    this.clear = function() {
+        token_list.children("li").each(function() {
+            if ($(this).children("input").length === 0) {
+                delete_token($(this));
+            }
+        });
+    }
+
+    this.add = function(item) {
+        add_token(item);
+    }
+
+    this.remove = function(item) {
+        token_list.children("li").each(function() {
+            if ($(this).children("input").length === 0) {
+                var currToken = $(this).data("tokeninput");
+                var match = true;
+                for (var prop in item) {
+                    if (item[prop] !== currToken[prop]) {
+                        match = false;
+                        break;
+                    }
+                }
+                if (match) {
+                    delete_token($(this));
+                }
+            }
+        });
+    }
 
     //
     // Private functions
     //
+
+    function checkTokenLimit() {
+        if(settings.tokenLimit !== null && token_count >= settings.tokenLimit) {
+            input_box.hide();
+            hide_dropdown();
+            return;
+        } else {
+            input_box.focus();
+        }
+    }
 
     function resize_input() {
         if(input_val === (input_val = input_box.val())) {return;}
@@ -381,8 +454,8 @@ $.TokenList = function (input, url_or_data, options) {
     }
 
     // Inner function to a token to the list
-    function insert_token(id, value) {
-        var this_token = $("<li><p>"+ value +"</p></li>")
+    function insert_token(item) {
+        var this_token = $("<li><p>"+ item.name +"</p></li>")
           .addClass(settings.classes.token)
           .insertBefore(input_token);
 
@@ -396,8 +469,8 @@ $.TokenList = function (input, url_or_data, options) {
             });
 
         // Store data on the token
-        var token_data = {"id": id, "name": value};
-        $.data(this_token.get(0), "tokeninput", token_data);
+        var token_data = {"id": item.id, "name": item.name};
+        $.data(this_token.get(0), "tokeninput", item);
 
         // Save this token for duplicate checking
         saved_tokens = saved_tokens.slice(0,selected_token_index).concat([token_data]).concat(saved_tokens.slice(selected_token_index));
@@ -419,7 +492,6 @@ $.TokenList = function (input, url_or_data, options) {
         if (settings.dontAdd)
             return;
 
-        var li_data = $.data(item.get(0), "tokeninput");
         var callback = settings.onAdd;
 
         // See if the token already exists and select it if we don't want duplicates
@@ -428,7 +500,7 @@ $.TokenList = function (input, url_or_data, options) {
             token_list.children().each(function () {
                 var existing_token = $(this);
                 var existing_data = $.data(existing_token.get(0), "tokeninput");
-                if(existing_data && existing_data.id === li_data.id) {
+                if(existing_data && existing_data.id === item.id) {
                     found_existing_token = existing_token;
                     return false;
                 }
@@ -442,14 +514,8 @@ $.TokenList = function (input, url_or_data, options) {
         }
 
         // Insert the new tokens
-        insert_token(li_data.id, li_data.name);
-
-        // Check the token limit
-        if(settings.tokenLimit !== null && token_count >= settings.tokenLimit) {
-            input_box.hide();
-            hide_dropdown();
-            return;
-        }
+        insert_token(item);
+        checkTokenLimit();
 
         // Clear input box
         input_box.val("");
@@ -459,7 +525,7 @@ $.TokenList = function (input, url_or_data, options) {
 
         // Execute the onAdd callback if defined
         if($.isFunction(callback)) {
-            callback(li_data);
+            callback.call(hidden_input,item);
         }
     }
 
@@ -540,7 +606,7 @@ $.TokenList = function (input, url_or_data, options) {
 
         // Execute the onDelete callback if defined
         if($.isFunction(callback)) {
-            callback(token_data);
+            callback.call(hidden_input,token_data);
         }
     }
 
@@ -613,7 +679,7 @@ $.TokenList = function (input, url_or_data, options) {
                     select_dropdown_item($(event.target).closest("li"));
                 })
                 .mousedown(function (event) {
-                    add_token($(event.target).closest("li"));
+                    add_token($(event.target).closest("li").data("tokeninput"));
                     input_box.blur();
                     return false;
                 })
@@ -633,7 +699,7 @@ $.TokenList = function (input, url_or_data, options) {
                     select_dropdown_item(this_li);
                 }
 
-                $.data(this_li.get(0), "tokeninput", {"id": value.id, "name": value.name});
+                $.data(this_li.get(0), "tokeninput", value);
             });
 
             show_dropdown();
@@ -738,7 +804,7 @@ $.TokenList = function (input, url_or_data, options) {
                 // Attach the success callback
                 ajax_params.success = function(results) {
                   if($.isFunction(settings.onResult)) {
-                      results = settings.onResult.call(this, results);
+                      results = settings.onResult.call(hidden_input, results);
                   }
                   cache.add(query, settings.jsonContainer ? results[settings.jsonContainer] : results);
 
@@ -756,6 +822,9 @@ $.TokenList = function (input, url_or_data, options) {
                     return row.name.toLowerCase().indexOf(query.toLowerCase()) > -1;
                 });
 
+                if($.isFunction(settings.onResult)) {
+                    results = settings.onResult.call(hidden_input, results);
+                }
                 cache.add(query, results);
                 populate_dropdown(query, results);
             }
